requirements.txt
google-generativeai>=0.3.0
pyaudio>=0.2.11
speechrecognition>=3.10.0
pydub>=0.25.1
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0

